+++ frontend/src/components/ScheduleAddButton/ScheduleAddButton.styles.ts
@@ -3,7 +3,6 @@ import { css, Theme } from '@emotion/react';
 const scheduleAddButton = ({ colors }: Theme) => css`
   width: 36px;
   height: 36px;
-
   border-radius: 50%;
 
   background: ${colors.YELLOW_500};
@@ -12,10 +11,6 @@ const scheduleAddButton = ({ colors }: Theme) => css`
   line-height: 20px;
   font-weight: 700;
   color: ${colors.WHITE};
-
-  &:hover {
-    opacity: 0.7;
-  }
 `;
 
 export { scheduleAddButton };
