+++ frontend/src/components/ScheduleAddButton/ScheduleAddButton.tsx
@@ -1,14 +1,13 @@
-/* eslint-disable import/named */
 import { useTheme } from '@emotion/react';
 
 import { scheduleAddButton } from './ScheduleAddButton.styles';
 
-import Button from '../@common/Button';
+import Button from '../@common/Button/Button';
 
 function ScheduleAddButton() {
   const theme = useTheme();
 
-  return <Button css={scheduleAddButton(theme)}>+</Button>;
+  return <Button cssProp={scheduleAddButton(theme)}>+</Button>;
 }
 
 export default ScheduleAddButton;
